feat(cli): add push command for service images
`push [--ignore-push-failures] [--tls-verify <bool>] [SERVICE...]` pushes each
service's image to its registry via the libpod /images/{name}/push endpoint.
Credentials come from Podman's auth file (podman login), so no auth handling is
needed. Services without an image are skipped; --ignore-push-failures continues
after a failure; --tls-verify false allows an insecure/local HTTP registry
(omitted leaves Podman's verify-on default). A mid-stream error line fails the
push.

Verified against a local registry:2 on Podman 5.4.2: push --tls-verify false
lands the image; an unreachable registry fails with a clear error.

main.rs reached the 500-line limit, so the diagnostic log formatter, tracing
init, internal-error notice, and help-framed CLI parse moved into a new startup
module.
