+++ internal/engine/mod.rs
@@ -5,7 +5,7 @@
 mod build;
 mod container;
 mod copy;
-pub use build::BuildOptions;
+pub use build::{BuildOptions, PushOptions};
 pub use lifecycle::RunOptions;
 pub use lock::ProjectLock;
 pub use query::{ExecOptions, ImagesOptions, PsOptions};
